requirements.txt

biopython
